Move private modules to utils
